implement rudimentary hibike_process based off of hibike_simulator and fix compilation issues in firmware

+++ hibike/lib/hibike/hibike_message.cpp
@@ -109,7 +109,7 @@ int send_data_update(uint16_t params) {
 
   for (uint16_t count = 0; (params >> count) > 0; count++) {
       if (params & (1<<count)){
-        int bytes_written = device_data_update(count, &msg.payload[msg.payload_length], msg.payload_length);
+        int bytes_written = device_data_update((uint8_t) count, &msg.payload[msg.payload_length], (size_t) msg.payload_length);
         if(bytes_written){
           msg.payload_length += bytes_written;
         }
